requirement.txt
pandas
httplib2
pyenchant
nltk
pattern
xlrd
openpyxl